refactor : useGetUser api 호출 훅 model -> api 로 이동
